% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-package.R
\docType{data}
\name{artifacts}
\alias{artifacts}
\title{Search for artifacts.}
\format{An object of class \code{NULL} of length 0.}
\usage{
artifacts
}
\description{
An entry point for artifact search. Exposes an interactive query
mechanism with tab-completion via the \code{$} (dollar) operator.
}
\keyword{datasets}
