% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-keys.R
\docType{data}
\name{DollarNamesMapping}
\alias{DollarNamesMapping}
\alias{createDollarNamesMapping}
\title{Assigned in .onLoad}
\format{An object of class \code{list} of length 1.}
\usage{
DollarNamesMapping

createDollarNamesMapping()
}
\description{
Assigned in .onLoad
}
\keyword{datasets}
