% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\docType{data}
\name{DollarNamesMapping}
\alias{DollarNamesMapping}
\title{Assigned in .onLoad}
\format{An object of class \code{NULL} of length 0.}
\usage{
DollarNamesMapping
}
\description{
Assigned in .onLoad
}
\keyword{datasets}
