% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking.R
\docType{data}
\name{state}
\alias{state}
\alias{initiate_state}
\alias{open_default_repo}
\alias{pick_branch}
\alias{start_tracking}
\alias{stop_tracking}
\title{Global state of the tracker.}
\format{An object of class \code{environment} of length 0.}
\usage{
state

initiate_state(state)

open_default_repo(state, env, create = FALSE)

pick_branch(repo, env)

start_tracking(state)

stop_tracking(state)
}
\arguments{
\item{state}{state object (environment).}

\item{env}{\code{environment} used to find the branch to attach to.}

\item{create}{if \code{TRUE}, create the repository if it does not exist.}

\item{repo}{a \code{\link[repository:repository]{repository::repository()}} object.}
}
\description{
\describe{
\item{repo}{a \code{\link[repository:repository]{repository::repository()}} object}
\item{task_callback_id}{id of the callback passed to \link{addTaskCallback}}
}

\code{initiate_state} assigns the default values to all
parameters of the global \code{state} object. By default it:
\itemize{
\item creates or open a file-system-based \link[repository:repository]{repository::repository}
\item turns tracking on
}
}
\keyword{datasets}
