% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tree}
\alias{print.tree}
\alias{print.history}
\alias{print.artifact}
\title{Pretty-print sets of artifacts.}
\usage{
\method{print}{tree}(x, ...)

\method{print}{history}(x, ...)

\method{print}{artifact}(x, ..., style = "full")
}
\arguments{
\item{x}{artifacts, e.g. returned  \link[repository:read_artifacts]{repository::read_artifacts}.}

\item{...}{Arguments passed on to \code{print}
\describe{
  \item{}{}
}}

\item{style}{\code{"full"}, \code{"short"} or \code{"line"}.}
}
\description{
Pretty-print sets of artifacts.

\code{print.explained} pretty-prints a description of an
artifact.
}
