% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_sample_repository}
\alias{use_sample_repository}
\title{Work with the sample repository.}
\usage{
use_sample_repository()
}
\description{
Switches the current repository to the sample one (see \code{\link[repository:sample_repository]{repository::sample_repository()}}
for more details). The sample repository is stored in a temporary
directory so as long as it is writeable, no changes are preserved
beyond the current R session.
}
