% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{new_specifier}
\alias{new_specifier}
\alias{is_specifier}
\alias{dollar_names.specifier}
\alias{dollar_name.specifier}
\title{Key-specifier object.}
\usage{
new_specifier(query, key)

is_specifier(x)

\method{dollar_names}{specifier}(x, pattern = "")

\method{dollar_name}{specifier}(x, i)
}
\description{
Key specifier serves in the process of building an object query. It
provides an interface to define the \code{tag} and its value that will be
sent to the repository when querying for objects.
}
\details{
Each key specifier must adhere to this rule: if the result of the
current query contains multiple objects, return a wrapped \code{query}
object. However, if there is only one object in the repository that
matches user query, return a wrapped single result.
}
