% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking.R
\name{create_state}
\alias{create_state}
\alias{is_state}
\alias{open_repo}
\alias{open_default_repo}
\alias{pick_branch}
\alias{start_tracking}
\alias{stop_tracking}
\title{Global state of the tracker.}
\usage{
create_state()

is_state(x)

open_repo(state, env, path, create)

open_default_repo(state, env, create = FALSE)

pick_branch(state, env)

start_tracking(state)

stop_tracking(state)
}
\arguments{
\item{x}{object being tested.}

\item{state}{state object (environment).}

\item{env}{\link{environment} used to find the branch to attach to.}

\item{path}{directory for the new/existing repository.}

\item{create}{if \code{TRUE}, create the repository if it does not exist.}
}
\description{
\code{create_state} creates a new \code{state} object and
assigns the default values to all its attributes.
}
\details{
\describe{
\item{repo}{a \code{\link[repository:repository]{repository::repository()}} object}
\item{task_callback_id}{id of the callback passed to \link{addTaskCallback}}
}
}
