% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking.R
\name{task_callback}
\alias{task_callback}
\title{A callback run after each top-level expression is evaluated.}
\usage{
task_callback(expr, result, successful, printed, state)
}
\arguments{
\item{expr}{Expression for the top-level task.}

\item{result}{Result of the top-level task.}

\item{successful}{A logical value indicating whether it was
successfully completed or not (always \code{TRUE} at present).}

\item{printed}{A logical value indicating whether the result was
printed or not.}

\item{state}{created with \link{new_state}, passed as the \code{data} argument
to \link{addTaskCallback}}
}
\value{
A logical value indicating whether to keep this function in
the list of active callbacks.
}
\description{
From \code{\link[=addTaskCallback]{addTaskCallback()}}: if the data argument was specified in
the call to addTaskCallback, that value is given as the fifth
argument.
}
