% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-package.R
\docType{package}
\name{ui}
\alias{ui}
\alias{ui-package}
\title{User Interface for the Repository of Artifacts}
\description{
Provides user interface for the \link{repository} package.

Package-level \link{options}:
\itemize{
\item \code{ui.attach_on_load} (default: \code{TRUE}) find the most up-to-date
\link{commit} in the repository and continue work from there
\item \code{ui.track} (default: \code{TRUE}) start tracking when the \code{ui} package
is loaded
}
}
