% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking.R
\name{create_state}
\alias{create_state}
\alias{initiate_state}
\alias{open_repo}
\alias{open_default_repo}
\alias{pick_branch}
\alias{start_tracking}
\alias{stop_tracking}
\title{Global state of the tracker.}
\usage{
create_state()

initiate_state(state)

open_repo(state, env, path, create)

open_default_repo(state, env, create = FALSE)

pick_branch(state, env)

start_tracking(state)

stop_tracking(state)
}
\arguments{
\item{state}{state object (environment).}

\item{env}{\link{environment} used to find the branch to attach to.}

\item{path}{directory for the new/existing repository.}

\item{create}{if \code{TRUE}, create the repository if it does not exist.}
}
\description{
\describe{
\item{repo}{a \code{\link[repository:repository]{repository::repository()}} object}
\item{task_callback_id}{id of the callback passed to \link{addTaskCallback}}
}

\code{initiate_state} assigns the default values to all
parameters of the global \code{state} object. By default it:
\itemize{
\item creates or open a file-system-based \link[repository:repository]{repository::repository}
\item turns tracking on
}
}
