% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking.R
\docType{data}
\name{state}
\alias{state}
\alias{initiate_state}
\alias{start_tracking}
\alias{stop_tracking}
\alias{open_repo}
\title{Global state of the tracker.}
\format{An object of class \code{environment} of length 0.}
\usage{
state

initiate_state()

start_tracking()

stop_tracking()

open_repo(create = FALSE)
}
\description{
\describe{
\item{repo}{a \link[repository:repositort]{repository::repositort} object}
\item{task_callback_id}{id of the callback passed to \link{addTaskCallback}}
}

\code{initiate_state} assigns the default values to all
parameters of the global \code{state} object. By default it:
\itemize{
\item creates or open a file-system-based \link[repository:repository]{repository::repository}
\item turns tracking on
}
}
\keyword{datasets}
