% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_london_meters}
\alias{use_london_meters}
\title{Work with the sample repository.}
\usage{
use_london_meters()
}
\description{
Switches the current repository to the sample one (see \code{\link[repository:london_meters]{repository::london_meters()}}
for more details). The sample repository is stored in a temporary
directory so as long as it is writeable, no changes are preserved
beyond the current R session.
}
