% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single-result.R
\name{single_result}
\alias{single_result}
\title{Wrapper for a result that contains a single artifact.}
\usage{
single_result(id, repo)
}
\arguments{
\item{id}{artifact identifier}
}
\description{
Wrapper for a result that contains a single artifact.
}
