% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking.R
\name{new_state}
\alias{new_state}
\alias{is_state}
\alias{state_reset}
\alias{open_repository}
\alias{pick_branch}
\alias{interactions}
\alias{start_tracking}
\alias{stop_tracking}
\title{Global state of the tracker.}
\usage{
new_state()

is_state(x)

state_reset(state)

open_repository(state, path, int = interactions())

pick_branch(state, env, int = interactions())

interactions(create_repository, clean_env, choose_commit, choose_branch)

start_tracking(state, callback_name)

stop_tracking(state)
}
\arguments{
\item{x}{object being tested.}

\item{state}{state object (environment).}

\item{path}{directory for the new/existing repository.}

\item{int}{an \link{interactions} object.}

\item{env}{\link{environment} used to find the branch to attach to.}
}
\description{
\code{new_state} creates a new \code{state} object and
assigns the default values to all its attributes.

\code{state_reset} assign default values to all attributes
of the \code{state} object.

\code{pick_branch} implements the logic of choosing the commit
to attach to.

\code{interactions} creates an object with a number of callbacks
to be used when decision can be delegated to the user. It provides default
implementations which either make the most straightforwards decision or
abort the call with a descriptive user message.

Provided callbacks:
\itemize{
\item \code{create_repository} to decide if a new repository should be created
\item \code{clean_env} to decide if objects in the global environment should be
removed in order to attach to repository
\item \code{choose_commit} when there are multiple commits matching global environment
\item \code{choose_branch} when there is more than one branch and global environment
is empty
}
}
\details{
\describe{
\item{repo}{a \code{\link[repository:repository]{repository::repository()}} object}
\item{task_callback_id}{id of the callback passed to \link{addTaskCallback}}
}
}
