% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{dollar_name.query}
\alias{dollar_name.query}
\title{Interactive UI for queries.}
\usage{
\method{dollar_name}{query}(x, n)
}
\arguments{
\item{x}{\link[repository:query]{repository::query} object.}

\item{n}{action name.}
}
\description{
UI for the \link[repository:query]{repository::query} object. It employs R's dollar operator
\code{$} to provide an interactive access to the repository of artifacts.
The \code{key} of the \code{$} operator denotes an action. Supported actions
are:
}
\details{
\itemize{
\item \code{history}
\itemize{
\item \code{tree}
\item \code{value}
\item query specifier: \code{name}, \code{id}, \code{class}, \code{time}, \code{session}
\item artifact name or identifier
}
}
}
