% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{wrapper}
\alias{wrapper}
\alias{wrap}
\alias{unwrap}
\alias{is_wrapper}
\alias{print.wrapper}
\alias{str.wrapper}
\alias{.DollarNames.wrapper}
\alias{$.wrapper}
\alias{[[.wrapper}
\alias{dollar_names}
\alias{dollar_name}
\alias{double_bracket}
\alias{dollar_names.wrapper}
\alias{dollar_name.wrapper}
\alias{double_bracket.wrapper}
\alias{dollar_names.default}
\alias{dollar_name.default}
\title{Wrapper and dollar-name interceptors.}
\usage{
wrap(x)

unwrap(x)

is_wrapper(x)

\method{print}{wrapper}(x, ...)

\method{str}{wrapper}(object, ...)

\method{.DollarNames}{wrapper}(x, pattern = "")

\method{$}{wrapper}(x, i)

\method{[[}{wrapper}(x, i)

dollar_names(x, pattern = "", ...)

dollar_name(x, i)

double_bracket(x, i)

\method{dollar_names}{wrapper}(x, pattern = "", ...)

\method{dollar_name}{wrapper}(x, i)

\method{double_bracket}{wrapper}(x, i)

\method{dollar_names}{default}(x, pattern = "", ...)

\method{dollar_name}{default}(x, i)
}
\arguments{
\item{x}{object to be wrapped, a \code{wrapper} object to be unwrapped
or printed, or any other object to be tested.}

\item{...}{further arguments passed to or from other methods.}

\item{object}{\code{wrapper} object.}

\item{pattern}{regular expression; only matching names are returned.}

\item{i}{key name, index value.}
}
\description{
In order to separate the user-facing API from the testable code that
implements user-facing API, all user-facing S3 classes are always
wrapped in the \code{"wrapper"} class. Moreover, instead of overloading
the \code{[[} and \code{$} operators and the \code{.DollarNames} method for all these
user-facing classes, three new methods are introduced, namely:
\code{double_bracket}, \code{dollar_name} and \code{dollar_names}, which have the
exact same semantics. This way, in the test code, accessing actual
data can be clearly distinguished from the user-facing tab-completion
mechanism.

\code{wrap} puts \code{x} in a list and sets that list's class
to \code{"wrapper"}.

\code{unwrap} returns the original wrapped object.

\code{print.wrapper} calls the actual \code{print} method for the
unwrapped object.

\code{str.wrapper} is required because RStudio calls str on
objects from the global environment, whose \code{.default} version calls
the \code{[[} operator, which produces a message. This is confusing to the
end user because they haven't requested an explicit object extraction.

the \code{.DollarNames} method and the \code{$} operator are the
only user-facing entry points into the tab-completion mechanism.

\code{dollar_names} is an equivalent of \code{.DollarNames} and
\code{dollar_name} is an equivalent of the \code{$} operator. All S3 classes
that are publicly exposed via the \code{"wrapper"} class need to define
their own version of these two methods to enable tab completion.
}
